requirements.txt
# FILE: requirements.txt

# Telegram Bot Framework
python-telegram-bot[ext]==21.2

# Core & Async
python-dotenv==1.0.1
pydantic-settings==2.3.4
asyncio==3.4.3

# Job Queue
celery==5.4.0
redis==5.0.7

# RAG & LLM
langchain==0.2.6
langchain-google-genai==1.0.6
google-genai==0.8.0
langchain-community==0.2.6
tavily-python==0.3.3  # <--- ADD THIS FOR SEARCH
# FIX: Downgrade to the version without the telemetry bug
chromadb==0.4.24
pypdf==4.2.0
unidecode==1.3.8
beautifulsoup4==4.12.3 # <--- ADD THIS FOR PARSING WEBPAGES

# Mind Map & File Handling
graphviz==0.20.3
aiohttp==3.9.5
aiofiles==23.2.1

# For user state management
filelock==3.15.4